requirements.txt
sentry-sdk[flask]
flask
black
python-dotenv
